Move utility functions dealing with resource paths (for BDD suites) to their own class
There are utility functions, used by the BDD suites, for finding (and adjusting)
resource paths, that have been living in `hapi.utils.keys.Ed25519Utils.java`.  They
clearly don't belong there.

I moved them to a separate class, `ResourceLocator`, still in `hapi-utils`.  Perhaps
they should actually be in a different module?

Fixes #6916
